Fix benchmark-lib
