+++ benchmark/lib/BenchmarkLib.hs
@@ -4,8 +4,9 @@ import Criterion.Main
 import qualified FunGraph
 import qualified FunGraph.Test
 import qualified Control.Monad.ST as ST
-import Data.Functor (void)
+import Data.Functor (void, (<&>))
 import Control.Monad ((<=<))
+import qualified Data.List.NonEmpty as NE
 
 testDataFileName :: FilePath
 testDataFileName = "data/all3.json"
@@ -22,16 +23,17 @@ main = do
       , bench "Thaw" $ nfAppIO (void . ST.stToIO . FunGraph.thaw) frozenGraph
       , bench "Thaw+freeze" $ nfAppIO (void . ST.stToIO . (FunGraph.freeze <=< FunGraph.thaw)) frozenGraph
       ]
-    , bgroup "Query"
-      [ bgroup "queryPaths" $
-          map (queryPaths mutGraph) FunGraph.Test.allTestCases
-      ]
+    , bgroup "Query" $
+        map (queryPaths mutGraph) FunGraph.Test.allTestCases
     ]
   where
     queryPaths mutGraph test =
-      let (maxCount, _) = FunGraph.Test.queryTest_args test
-      in bench (FunGraph.Test.queryTest_name test <> " maxCount=" <> show maxCount) $
-        nfAppIO (\g -> ST.stToIO $ FunGraph.runGraphAction g $ FunGraph.Test.queryTest_runQuery test) mutGraph
+      let args@(maxCount, _) = FunGraph.Test.queryTest_args test
+          blahName = FunGraph.Test.queryTest_name test <> " maxCount=" <> show maxCount
+      in bgroup blahName $ NE.toList $
+        FunGraph.Test.mkQueryFunctions False <&> \(queryFunctionName, queryFunction) ->
+          bench queryFunctionName $
+            nfAppIO (queryFunction args) mutGraph
 
     readGraphData fileName =
       either fail pure =<< FunGraph.fileReadDeclarationMap fileName
